+++ examples/private-service-access/variables.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2026 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,7 @@
  * limitations under the License.
  */
 
-terraform {
-  required_version = ">= 1.3"
-
-  required_providers {
-    google-beta = {
-      source  = "hashicorp/google-beta"
-      version = ">= 3.0, < 8"
-    }
-  }
-
-  provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:routes-beta/v18.1.2"
-  }
+variable "project_id" {
+  description = "The ID of the project where the network will be created."
+  type        = string
 }