+++ examples/routes/main.tf
@@ -23,7 +23,7 @@
 # [START vpc_static_route_create]
 module "google_compute_route" {
   source       = "../../modules/routes" #adding local path
-  project_id   = var.project_id # Replace this with your project ID in quotes
+  project_id   = var.project_id         # Replace this with your project ID in quotes
   network_name = "default"
 
   routes = [