+++ main.tf
@@ -90,3 +90,13 @@ module "firewall_rules" {
   ingress_rules = var.ingress_rules
   egress_rules  = var.egress_rules
 }
+
+module "private_service_access" {
+  source = "./modules/private-service-access"
+  count  = var.private_service_access_config.enable_private_services_connection ? 1 : 0
+
+  project_id    = var.project_id
+  network_id    = module.vpc.network_id
+  address_name  = var.private_service_access_config.address_name
+  prefix_length = var.private_service_access_config.prefix_length
+}