+++ variables.tf
@@ -60,12 +60,14 @@ variable "subnets" {
     role                             = optional(string)
     stack_type                       = optional(string)
     ipv6_access_type                 = optional(string)
+    ip_collection                    = optional(string)
+    external_ipv6_prefix             = optional(string)
   }))
   description = "The list of subnets being created"
 }
 
 variable "secondary_ranges" {
-  type        = map(list(object({ range_name = string, ip_cidr_range = string })))
+  type        = map(list(object({ range_name = string, ip_cidr_range = optional(string), reserved_internal_range = optional(string) })))
   description = "Secondary ranges that will be used in some of the subnets"
   default     = {}
 }
@@ -248,3 +250,18 @@ variable "bgp_inter_region_cost" {
   description = "Specifies the BGP inter-region cost mode. Valid values are `DEFAULT` or `ADD_COST_TO_MED`."
   default     = null
 }
+
+variable "private_service_access_config" {
+  description = "Configuration for Private Service Access (PSA) connection."
+  type = object({
+    enable_private_services_connection = bool
+    address_name                       = string
+    prefix_length                      = number
+  })
+  default = {
+    enable_private_services_connection = false
+    address_name                       = "private-ip-address"
+    prefix_length                      = 16
+  }
+}
+